refactor(websites): eliminar env vars hardcodeadas de gc-belleza
Reemplaza NEXT_PUBLIC_TENANT_SLUG por getClientTenantSlug() y
NEXT_PUBLIC_CONTACT_* por useTenantContact() del TenantContext.
Los datos de contacto y el slug ahora son dinámicos por tenant.

Closes #234
