+++ frontend/src/components/common/LocationSection.tsx
@@ -11,8 +11,8 @@ export function LocationSection() {
   const contact = useTenantContact();
 
   // Build address string from contact info
-  const address = contact?.address || process.env.NEXT_PUBLIC_CONTACT_ADDRESS || 'Calle Principal 123, Madrid';
-  const phone = contact?.phone || process.env.NEXT_PUBLIC_CONTACT_PHONE || '+34 912 345 678';
+  const address = contact?.address;
+  const phone = contact?.phone;
   const schedule = 'L-V 9:00-19:00 · Sáb 10:00-14:00';
   return (
     <section className="py-12 bg-muted/30">
@@ -20,16 +20,20 @@ export function LocationSection() {
         <div className="flex flex-col md:flex-row items-center justify-between gap-6">
           {/* Info compacta */}
           <div className="flex flex-wrap items-center justify-center md:justify-start gap-x-8 gap-y-3">
-            <div className="flex items-center gap-2 text-sm text-foreground/80">
-              <MapPin className="h-4 w-4 text-primary shrink-0" />
-              <span>{address}</span>
-            </div>
-            <div className="flex items-center gap-2 text-sm text-foreground/80">
-              <Phone className="h-4 w-4 text-primary shrink-0" />
-              <a href={`tel:${phone}`} className="hover:text-primary transition-colors">
-                {phone}
-              </a>
-            </div>
+            {address && (
+              <div className="flex items-center gap-2 text-sm text-foreground/80">
+                <MapPin className="h-4 w-4 text-primary shrink-0" />
+                <span>{address}</span>
+              </div>
+            )}
+            {phone && (
+              <div className="flex items-center gap-2 text-sm text-foreground/80">
+                <Phone className="h-4 w-4 text-primary shrink-0" />
+                <a href={`tel:${phone}`} className="hover:text-primary transition-colors">
+                  {phone}
+                </a>
+              </div>
+            )}
             <div className="flex items-center gap-2 text-sm text-foreground/80">
               <Clock className="h-4 w-4 text-primary shrink-0" />
               <span>{schedule}</span>
